therapieland/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
fhir_core==1.1.4
fhir.resources==8.0.0
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
drf-spectacular==0.28.0